added fontawesome cdn, added fontfamily, refactored navbar and footer
