+++ src/components/Footer/Footer.jsx
@@ -1,5 +1,5 @@
 import './footer.css'
-import WhiteLogo from '../../assets/logo-white.png'
+import WhiteLogo from '../../assets/labable-white.svg'
 import InstaLogo from '../../assets/insta.png'
 import TeamsLogo from '../../assets/teams.png'
 import FbLogo from '../../assets/facebook.png'
@@ -7,14 +7,17 @@ import FbLogo from '../../assets/facebook.png'
 export default function Footer(){
     return(
         <>
-            <div className="footer-container">
+            <footer className="footer-container">
                 <div className="summary">
                     <div className="business">
-                        <img src={WhiteLogo} alt="" />
+                        <div className="white-logo">
+                            <img src={WhiteLogo} alt="Labable" className='logo-white'/>
+                            <h1><span>Laba</span><span className='highlight-tag'>ble</span></h1>
+                        </div>
                         <p>Smart Laundry Management System for modern living</p>
                     </div>
                     <div className="quick-links">
-                        <p className='footer-title'>Quick Links</p>
+                        <h2 className='footer-title'>Quick Links</h2>
                         <p>Home</p>
                         <p>About</p>
                         <p>Services</p>
@@ -22,7 +25,7 @@ export default function Footer(){
                         <p>Create Order</p>
                     </div>
                     <div className="services">
-                        <p className='footer-title'>Services</p>
+                        <h2 className='footer-title'>Services</h2>
                         <p>Washing</p>
                         <p>Dry Cleaning</p>
                         <p>Iron & Press</p>
@@ -30,16 +33,16 @@ export default function Footer(){
                         <p>Pickup & Delivery</p>
                     </div>
                     <div className="find-us-at">
-                        <p className='footer-title'>Find Us At</p>
+                        <h2 className='footer-title'>Find Us At</h2>
                         <div className="socials">
-                            <img src={InstaLogo} alt="" />
-                            <img src={TeamsLogo} alt="" />
-                            <img src={FbLogo} alt="" />
+                            <i className='fa-brands fa-square-instagram'></i>
+                            <i className='fa-brands fa-square-twitter'></i>
+                            <i className='fa-brands fa-square-facebook'></i>
                         </div>
                     </div>
                 </div>
                 <p className='copyright'>@2025 Labable. All right reserved. BSIT 3G - G1 Group 4</p>
-            </div>
+            </footer>
         </>
     )
 }
