+++ src/components/Footer/footer.css
@@ -1,6 +1,3 @@
-.footer-container *{
-    color: white;
-}
 .footer-container{
     background-color: black;
     padding: 50px 0;
@@ -8,37 +5,74 @@
     flex-direction: column;
     gap: 50px;
     width: 100%;
-}
-
-.summary p{
-    margin: 0;
+    color: var(--fg-light);
 }
 
 .summary{
     display: flex;
     flex-direction: row;
+    flex-wrap: wrap;
     justify-content: space-between;
-    padding-bottom: 20px;
-    border-bottom: 1px solid white;
-}
-
-.summary div{
-    width: 450px;
-    padding: 0px 50px;
+    padding: 20px 0;
+    border-bottom: 1px solid var(--fg-light);
+    gap: 20px;
 }
 
 .footer-title{
-    font-size: 30px;
+    font-size: 2rem;
     font-weight: bold;
-    padding-bottom: 8px;
 }
 
-.find-us-at .socials{
-    width: 100%;
+.business, .quick-links, .services, .find-us-at{
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+    flex-basis: 0;
+    padding: 0px 50px;
+}
+
+.white-logo{
     display: flex;
     flex-direction: row;
-    justify-content: space-between;
-    padding: 0px;
+    align-items: start;
+    gap: 10px;
+}
+
+.white-logo img{
+    width: 50px;
+    height: 50px;
+}
+
+.white-logo h1{
+    color: var(--fg-light);
+}
+
+.white-logo h1 span{
+    font-weight: bold;
+    font-size: 2rem;
+}
+
+.white-logo .highlight-tag{
+    color: var(--sf-dark);
+}
+
+.socials{
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+}
+
+.socials i{
+    font-size: 2rem;
+    cursor: pointer;
+    border-radius: 10px;
+    padding: 5px;
+}
+
+.socials i:hover{
+    transition: ease-in-out 0.3s;
+    color: var(--fg-dark);
+    background-color: var(--bg-light);
 }
 
 .copyright{
