+++ src/components/NavBar/nav-bar.css
@@ -8,9 +8,12 @@
     justify-content: space-between;
     align-items: center;
     width: 100%;
-    padding: 10px 20px;
-    color: rgba(255, 255, 255, 0.95);
-    border: 0.5px solid rgba(0, 0, 0, 0.3);
+    height: 80px;
+    padding: 10px 30px;
+    border-bottom: 0.5px solid var(--sf-dark);
+    position: fixed;
+    z-index: 1000;
+    background-color: color-mix(in srgb, var(--bg-light) 95%, transparent);
 }
 
 .links{
@@ -21,7 +24,8 @@
 
 .nav-link{
     text-decoration: none;
-    color: black;
+    color: var(--fg-dark);
+    font-size: 1rem;
 }
 
 .login-signup{
@@ -37,18 +41,43 @@
 }
 
 .login-signup .signup{
-    color: white;
-    background-color: black;
+    color: var(--fg-light);
+    background-color: var(--fg-dark);
     border: 1px solid;
     transition: ease 0.3s;
 }
 
 .login-signup .signup:hover{
-    color: black;
-    background-color: white;
+    color: var(--fg-dark);
+    background-color: var(--bg-light);
 }
 
 .login-signup .login:hover{
     transition: ease 0.3s;
-    background-color: rgba(91, 91, 91, .20);
+    background-color: color-mix(in srgb, var(--sf-dark) 20%, transparent);
+}
+
+.logo{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 10px;
+}
+
+.logo img{
+    width: 50px;
+    height: 50px;
+}
+
+.logo h1{
+    color: var(--fg-dark);
+}
+
+.logo h1 span{
+    font-weight: bold;
+    font-size: 2rem;
+}
+
+.logo .highlight-tag{
+    color: var(--sf-dark);
 }
