+++ src/pages/Home/Home.jsx
@@ -1,8 +1,8 @@
 import './home.css'
 export default function Home(){
     return(
-        <>
+        <section className='homepage-container'>
             <h1>Home</h1>
-        </>
+        </section>
     )
 }
