+++ src/pages/Home/home.css
@@ -1,0 +1,3 @@
+.homepage-container{
+    margin-top: 100px;
+}
